+++ GetArchive-webext/build.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-web-ext build --ignore-files build.sh images *.md
+web-ext build --ignore-files build.sh images *.md --overwrite-dest